+++ src/main/groovy/scorekeeper/sql/AutomaticSQLMetricsActor.java
@@ -21,14 +21,14 @@ public class AutomaticSQLMetricsActor extends SQLMetricsActor {
 		
 		int i = 0;
 		for (String metricName: metric.getMetricNames()){
-			int scalar = rs.getInt(++i);
+			double scalar = rs.getDouble(++i);
 			writeCounter(scalar, metricName);
 		}
 		
 		rs.close();
 	}
 
-	protected void writeCounter(int counter, String metricName) {
+	protected void writeCounter(double counter, String metricName) {
 		statsClient.gauge(metricName, counter);
 		System.out.print("s");
 //		System.out.println(metricName + ":" + counter);
